requirement.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
openpyxl==2.6.2
Pillow==6.0.0
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
openpyxl==2.6.2
Pillow==6.0.0
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
openpyxl==2.6.2
Pillow==6.0.0
pymongo==3.8.0
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
openpyxl==2.6.2
Pillow==6.0.0
pymongo==3.8.0
python-dotenv==0.10.1
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
